Rename AutoscalerFlinkMetrics to AutoscalerMetrics to avoid confusion
